seems to be a correct version now

+++ util.h
@@ -18,7 +18,7 @@ void fatal(char *msg);
 void threshold_vector(float *vec, int n, double T);
 
 float *readfeats_file(char *fn, int D, int *fA, int *fB, int *N);
-short *readsignal_file(char *fn, int *N);
+float *readsignal_file(char *fn, int *N);
 void writefeats_file(char *fn, float *feats, int D, int nframes);
 void printfeats_file(char *fn, float *feats, int D, int nframes);
 char *mmap_file(char *fn, int *len);
